build(sys): turn off unneeded regex features
regex is only needed as a build dependency, so we don't need the "perf"
feature set.

We're only using it to match version numbers, so we don't need most of
the "unicode" feature, either.

For the root hdf5 crate, it's only needed as a dev dependency, which
needs even less.